msm: ipa: driver upgrade for new IPAv2.5 core
Driver code changes to work with new IPAv2.5 core.
Changes include:
- registers interface changes (base offset
  change, fields offset change and addition of new registers)
- support higher HOLB timeout value (32 bit).

+++ include/linux/ipa.h
@@ -253,10 +253,12 @@ struct ipa_ep_cfg_route {
  * struct ipa_ep_cfg_holb - head of line blocking configuration in IPA end-point
  * @en: enable(1 => ok to drop pkt)/disable(0 => never drop pkt)
  * @tmr_val: duration in units of 128 IPA clk clock cyles [0,511], 1 clk=1.28us
+ *	     IPAv2.5 support 32 bit HOLB timeout value, previous versions
+ *	     supports 16 bit
  */
 struct ipa_ep_cfg_holb {
 	u16 en;
-	u16 tmr_val;
+	u32 tmr_val;
 };
 
 /**
